Move the implementation of check_numerics to paddle.amp.
